test: refactor destroy controller test away from gocheck
