+++ cmd/jaas/cmd/destroycontroller.go
